Refactor initial transfer request structure to assist elimination
